+++ e2e/tests/sds_node_configurator_test.go
@@ -21,6 +21,8 @@ import (
 	"crypto/md5"
 	"crypto/sha1"
 	"encoding/hex"
+	"encoding/json"
+	"errors"
 	"fmt"
 	"math/rand"
 	"os"
@@ -34,10 +36,14 @@ import (
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
+	virtv1alpha2 "github.com/deckhouse/virtualization/api/core/v1alpha2"
 	corev1 "k8s.io/api/core/v1"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	k8sruntime "k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/client-go/dynamic"
+	k8sclient "k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/client-go/rest"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -62,7 +68,16 @@ const (
 	// Direct SSH to nodes for lsblk can hit transient "handshake failed: EOF" (sshd/network) after heavy I/O.
 	e2eLsblkSSHMaxRetries   = 6
 	e2eLsblkSSHRetryInterval = 15 * time.Second
+
+	// alwaysCreateNew: aligned with github.com/deckhouse/storage-e2e/internal/config
+	e2eClusterCreationTimeout = 90 * time.Minute
+	e2eModuleDeployTimeout    = 15 * time.Minute
 )
+
+// clusterResumeState mirrors storage-e2e cluster-state.json (namespace after VMs are created).
+type clusterResumeState struct {
+	Namespace string `json:"namespace"`
+}
 
 func e2eConfigNamespace() string {
 	if v := os.Getenv("TEST_CLUSTER_NAMESPACE"); v != "" {
@@ -324,12 +339,11 @@ var _ = Describe("Sds Node Configurator", Ordered, func() {
 	})
 
 	// ---=== TEST CLUSTER IS CREATED AND READY HERE ===--- //
-	// For alwaysCreateNew we call CreateTestCluster + WaitForTestClusterReady explicitly so on failure we delete
-	// the test namespace on the base cluster (VMs) before the spec ends. Other modes use storage-e2e as-is.
+	// alwaysCreateNew: CreateTestCluster + WaitForTestClusterReady; on failure we clean base cluster (see createE2EAlwaysNewClusterWithCleanupOnFailure).
 
 	It("should create test cluster", func() {
 		if e2eConfigTestClusterCreateMode() == "alwaysCreateNew" {
-			testClusterResources = createE2EAlwaysNewClusterWithSyncCleanupOnFailure()
+			testClusterResources = createE2EAlwaysNewClusterWithCleanupOnFailure()
 			return
 		}
 		if e2eConfigTestClusterCreateMode() == "alwaysUseExisting" {
@@ -1246,4 +1260,221 @@ func e2eClusterStateJSONPath() (string, error) {
 	}
 	base := strings.TrimSuffix(filepath.Base(file), filepath.Ext(file))
 	return filepath.Join(repoRoot, "temp", base, "cluster-state.json"), nil
+}
+
+func e2eNewVirtClient(cfg *rest.Config) (client.Client, error) {
+	sch := k8sruntime.NewScheme()
+	if err := virtv1alpha2.SchemeBuilder.AddToScheme(sch); err != nil {
+		return nil, err
+	}
+	return client.New(cfg, client.Options{Scheme: sch})
+}
+
+// e2eCleanupBaseClusterNamespaceWorkload deletes namespaced virtualization objects on the base cluster in order:
+// VirtualMachineBlockDeviceAttachment → VirtualDisk → VirtualMachine, then the namespace (Deckhouse DVP).
+func e2eCleanupBaseClusterNamespaceWorkload(ctx context.Context, cfg *rest.Config, ns string) {
+	cl, err := e2eNewVirtClient(cfg)
+	if err != nil {
+		GinkgoWriter.Printf("    ⚠️  cleanup: virt client: %v (will still try namespace delete)\n", err)
+		e2eDeleteNamespaceBestEffort(ctx, cfg, ns)
+		return
+	}
+
+	var vmbdaList virtv1alpha2.VirtualMachineBlockDeviceAttachmentList
+	if err := cl.List(ctx, &vmbdaList, client.InNamespace(ns)); err != nil {
+		GinkgoWriter.Printf("    ⚠️  cleanup: list VMBDA: %v\n", err)
+	} else {
+		for i := range vmbdaList.Items {
+			if err := cl.Delete(ctx, &vmbdaList.Items[i]); err != nil && !apierrors.IsNotFound(err) {
+				GinkgoWriter.Printf("    ⚠️  cleanup: delete VMBDA %s: %v\n", vmbdaList.Items[i].Name, err)
+			}
+		}
+	}
+
+	var vdList virtv1alpha2.VirtualDiskList
+	if err := cl.List(ctx, &vdList, client.InNamespace(ns)); err != nil {
+		GinkgoWriter.Printf("    ⚠️  cleanup: list VirtualDisk: %v\n", err)
+	} else {
+		for i := range vdList.Items {
+			if err := cl.Delete(ctx, &vdList.Items[i]); err != nil && !apierrors.IsNotFound(err) {
+				GinkgoWriter.Printf("    ⚠️  cleanup: delete VirtualDisk %s: %v\n", vdList.Items[i].Name, err)
+			}
+		}
+	}
+
+	var vmList virtv1alpha2.VirtualMachineList
+	if err := cl.List(ctx, &vmList, client.InNamespace(ns)); err != nil {
+		GinkgoWriter.Printf("    ⚠️  cleanup: list VirtualMachine: %v\n", err)
+	} else {
+		for i := range vmList.Items {
+			if err := cl.Delete(ctx, &vmList.Items[i]); err != nil && !apierrors.IsNotFound(err) {
+				GinkgoWriter.Printf("    ⚠️  cleanup: delete VirtualMachine %s: %v\n", vmList.Items[i].Name, err)
+			}
+		}
+	}
+
+	e2eDeleteNamespaceBestEffort(ctx, cfg, ns)
+}
+
+func e2eDeleteNamespaceBestEffort(ctx context.Context, cfg *rest.Config, ns string) {
+	cs, err := k8sclient.NewForConfig(cfg)
+	if err != nil {
+		GinkgoWriter.Printf("    ⚠️  cleanup: kubernetes client: %v\n", err)
+		return
+	}
+	delCtx, cancel := context.WithTimeout(ctx, e2eClusterCleanupTimeout)
+	defer cancel()
+	err = cs.CoreV1().Namespaces().Delete(delCtx, ns, metav1.DeleteOptions{})
+	if apierrors.IsNotFound(err) {
+		GinkgoWriter.Printf("    ℹ️  cleanup: namespace %q already gone\n", ns)
+		return
+	}
+	if err != nil {
+		GinkgoWriter.Printf("    ⚠️  cleanup: delete namespace %q: %v\n", ns, err)
+		return
+	}
+	GinkgoWriter.Printf("    ✅ cleanup: namespace %q deletion submitted\n", ns)
+}
+
+func kubeconfigDirForNamespaceDelete(clusterStatePath string, _ error) (dir string, tmpToRemove string, err error) {
+	if clusterStatePath != "" {
+		d := filepath.Dir(clusterStatePath)
+		if mkErr := os.MkdirAll(d, 0755); mkErr != nil {
+			return "", "", mkErr
+		}
+		return d, "", nil
+	}
+	d, mkErr := os.MkdirTemp("", "e2e-ns-del-kube-*")
+	if mkErr != nil {
+		return "", "", mkErr
+	}
+	return d, d, nil
+}
+
+// e2eSyncCleanupBaseClusterNamespace connects to the base cluster, removes VMBDA/VD/VM in namespace, then deletes the namespace.
+// Used when alwaysCreateNew fails mid-run so resources are not left until AfterAll.
+func e2eSyncCleanupBaseClusterNamespace(ctx context.Context, clusterStatePath string, statePathErr error) {
+	kubeconfigDir, cleanupDir, err := kubeconfigDirForNamespaceDelete(clusterStatePath, statePathErr)
+	if err != nil {
+		GinkgoWriter.Printf("    ⚠️  cleanup: %v\n", err)
+		return
+	}
+	if cleanupDir != "" {
+		defer func() { _ = os.RemoveAll(cleanupDir) }()
+	}
+
+	var ns string
+	if clusterStatePath != "" {
+		data, readErr := os.ReadFile(clusterStatePath)
+		if readErr == nil {
+			var st clusterResumeState
+			if json.Unmarshal(data, &st) == nil && st.Namespace != "" {
+				ns = st.Namespace
+			}
+		} else if !errors.Is(readErr, os.ErrNotExist) {
+			GinkgoWriter.Printf("    ⚠️  cleanup: read %s: %v\n", clusterStatePath, readErr)
+		}
+	}
+	if ns == "" {
+		ns = e2eConfigNamespace()
+		GinkgoWriter.Printf("    ▶️  cleanup: namespace from TEST_CLUSTER_NAMESPACE=%q\n", ns)
+	} else {
+		GinkgoWriter.Printf("    ▶️  cleanup: namespace from cluster-state.json: %q\n", ns)
+	}
+
+	sshHost := e2eConfigSSHHost()
+	sshUser := e2eConfigSSHUser()
+	if sshHost == "" || sshUser == "" {
+		GinkgoWriter.Printf("    ⚠️  cleanup: SSH_HOST or SSH_USER not set, skip base cluster cleanup\n")
+		return
+	}
+
+	sshKeyPath, err := cluster.GetSSHPrivateKeyPath()
+	if err != nil {
+		GinkgoWriter.Printf("    ⚠️  cleanup: SSH key: %v\n", err)
+		return
+	}
+
+	useJump := e2eConfigSSHJumpHost() != ""
+	var opts cluster.ConnectClusterOptions
+	if !useJump {
+		opts = cluster.ConnectClusterOptions{
+			SSHUser: sshUser, SSHHost: sshHost, SSHKeyPath: sshKeyPath,
+			UseJumpHost:         false,
+			KubeconfigOutputDir: kubeconfigDir,
+		}
+	} else {
+		jumpUser := e2eConfigSSHJumpUser()
+		if jumpUser == "" {
+			jumpUser = sshUser
+		}
+		jumpKey := e2eConfigSSHJumpKeyPath()
+		if jumpKey == "" {
+			jumpKey = sshKeyPath
+		}
+		opts = cluster.ConnectClusterOptions{
+			SSHUser: jumpUser, SSHHost: e2eConfigSSHJumpHost(), SSHKeyPath: jumpKey,
+			UseJumpHost: true, TargetUser: sshUser, TargetHost: sshHost, TargetKeyPath: sshKeyPath,
+			KubeconfigOutputDir: kubeconfigDir,
+		}
+	}
+
+	baseRes, err := cluster.ConnectToCluster(ctx, opts)
+	if err != nil {
+		GinkgoWriter.Printf("    ⚠️  cleanup: connect to base cluster: %v\n", err)
+		return
+	}
+	defer func() {
+		if baseRes.TunnelInfo != nil && baseRes.TunnelInfo.StopFunc != nil {
+			_ = baseRes.TunnelInfo.StopFunc()
+		}
+		if baseRes.SSHClient != nil {
+			_ = baseRes.SSHClient.Close()
+		}
+	}()
+
+	GinkgoWriter.Printf("    ▶️  cleanup: ordered teardown VMBDA → VirtualDisk → VirtualMachine → Namespace %q\n", ns)
+	e2eCleanupBaseClusterNamespaceWorkload(ctx, baseRes.Kubeconfig, ns)
+
+	if clusterStatePath != "" {
+		_ = os.Remove(clusterStatePath)
+	}
+}
+
+// createE2EAlwaysNewClusterWithCleanupOnFailure runs CreateTestCluster + WaitForTestClusterReady for alwaysCreateNew.
+// On failure, cleans the base cluster namespace (VD/VM/ns) inside the same spec before Gomega fails.
+func createE2EAlwaysNewClusterWithCleanupOnFailure() *cluster.TestClusterResources {
+	statePath, statePathErr := e2eClusterStateJSONPath()
+
+	yamlName := os.Getenv("YAML_CONFIG_FILENAME")
+	if yamlName == "" {
+		yamlName = "cluster_config.yml"
+	}
+
+	createCtx, cancel := context.WithTimeout(context.Background(), e2eClusterCreationTimeout)
+	defer cancel()
+
+	res, err := cluster.CreateTestCluster(createCtx, yamlName)
+	if err != nil {
+		GinkgoWriter.Printf("    ▶️  CreateTestCluster failed; cleaning base cluster namespace before spec exits...\n")
+		e2eSyncCleanupBaseClusterNamespace(context.Background(), statePath, statePathErr)
+		Expect(err).NotTo(HaveOccurred(), "Test cluster should be created successfully")
+	}
+
+	waitCtx, cancel2 := context.WithTimeout(context.Background(), e2eModuleDeployTimeout)
+	defer cancel2()
+	err = cluster.WaitForTestClusterReady(waitCtx, res)
+	if err != nil {
+		GinkgoWriter.Printf("    ▶️  WaitForTestClusterReady failed; cleaning base cluster then storage-e2e cleanup...\n")
+		e2eSyncCleanupBaseClusterNamespace(context.Background(), statePath, statePathErr)
+		cctx, ccancel := context.WithTimeout(context.Background(), e2eClusterCleanupTimeout)
+		cleanupErr := cluster.CleanupTestCluster(cctx, res)
+		ccancel()
+		if cleanupErr != nil {
+			GinkgoWriter.Printf("    ⚠️  CleanupTestCluster after failed readiness: %v\n", cleanupErr)
+		}
+		Expect(err).NotTo(HaveOccurred(), "Test cluster should become ready")
+	}
+
+	return res
 }
